+++ torchtune/datasets/_grammar.py
@@ -10,7 +10,7 @@ from typing import Any, Callable, Dict, Optional, Union
 from torchtune.data import InputOutputToMessages
 from torchtune.datasets._packed import PackedDataset
 from torchtune.datasets._sft import SFTDataset
-from torchtune.modules.tokenizers import ModelTokenizer
+from torchtune.modules.transforms.tokenizers import ModelTokenizer
 
 
 def grammar_dataset(
