+++ torchtune/datasets/_wikitext.py
@@ -13,7 +13,7 @@ from torchtune.datasets._text_completion import (
     TextCompletionDataset,
 )
 
-from torchtune.modules.tokenizers import ModelTokenizer
+from torchtune.modules.transforms.tokenizers import ModelTokenizer
 
 
 def wikitext_dataset(
