+++ torchtune/models/llama3/_model_builders.py
@@ -13,7 +13,7 @@ from torchtune.models.llama3._tokenizer import Llama3Tokenizer
 
 from torchtune.modules import TransformerDecoder
 from torchtune.modules.peft import LORA_ATTN_MODULES
-from torchtune.modules.tokenizers import parse_hf_tokenizer_json
+from torchtune.modules.transforms.tokenizers import parse_hf_tokenizer_json
 
 
 """
